requirements.txt
aio-pika==9.3.1
aiormq==6.7.7
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.1.2
click==8.1.8
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.19.1
email-validator==2.1.0
exceptiongroup==1.3.0
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
httpx==0.25.2
idna==3.10
loguru==0.7.3
multidict==6.0.5
pamqp==3.2.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.4.8
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.0.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==0.21.1
python-jose==3.4.0
redis==5.0.1
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.22.0
yarl==1.9.4
